[ci skip] refactor: create request context from network context
