+++ brightray/BUILD.gn
@@ -36,8 +36,6 @@ static_library("brightray") {
     "browser/brightray_paths.h",
     "browser/browser_client.cc",
     "browser/browser_client.h",
-    "browser/browser_context.cc",
-    "browser/browser_context.h",
     "browser/browser_main_parts.cc",
     "browser/browser_main_parts.h",
     "browser/browser_main_parts_mac.mm",
