+++ brightray/browser/browser_client.cc
@@ -7,7 +7,6 @@
 #include "base/lazy_instance.h"
 #include "base/no_destructor.h"
 #include "base/path_service.h"
-#include "brightray/browser/browser_context.h"
 #include "brightray/browser/browser_main_parts.h"
 #include "brightray/browser/devtools_manager_delegate.h"
 #include "brightray/browser/media/media_capture_devices_dispatcher.h"
