+++ filenames.gni
@@ -303,6 +303,8 @@ filenames = {
     "atom/browser/net/url_request_string_job.h",
     "atom/browser/net/url_request_buffer_job.cc",
     "atom/browser/net/url_request_buffer_job.h",
+    "atom/browser/net/url_request_context_getter.cc",
+    "atom/browser/net/url_request_context_getter.h",
     "atom/browser/net/url_request_fetch_job.cc",
     "atom/browser/net/url_request_fetch_job.h",
     "atom/browser/net/url_request_stream_job.cc",
@@ -316,8 +318,6 @@ filenames = {
     "atom/browser/relauncher.h",
     "atom/browser/render_process_preferences.cc",
     "atom/browser/render_process_preferences.h",
-    "atom/browser/request_context_delegate.cc",
-    "atom/browser/request_context_delegate.h",
     "atom/browser/session_preferences.cc",
     "atom/browser/session_preferences.h",
     "atom/browser/special_storage_policy.cc",
